+++ web/app/themes/rebe/template-how-to.php
@@ -1,0 +1,14 @@
+<?php
+
+/*
+
+    Template Name: How-to Page
+
+    This template is used for How-to page content that has a unique layout.
+
+ */
+
+?>
+<main class="o-row">
+  <?php get_template_part('templates/content', 'how-to'); ?>
+</main>